+++ packages/react-shared-components/src/context.ts
@@ -1,6 +1,5 @@
 import { createContext } from 'react'
-import { IArrayListProps, PreviewTextConfigProps, ILayoutProps } from './types'
+import { IArrayListProps, PreviewTextConfigProps } from './types'
 
 export const PreviewTextContext = createContext<PreviewTextConfigProps>({})
 export const ArrayContext = createContext<IArrayListProps>({})
-export const LayoutContext = createContext<Omit<ILayoutProps, 'children'>>({})
