+++ packages/react-shared-components/src/index.ts
@@ -3,4 +3,3 @@ export * from './PreviewText'
 export * from './PasswordStrength'
 export * from './types'
 export * from './DragListView'
-export * from './Layout'
