+++ packages/react/src/index.ts
@@ -23,9 +23,11 @@ export * from './components/VirtualField'
 export * from './components/FormSpy'
 export * from './components/FormProvider'
 export * from './components/FormConsumer'
+export * from './components/Layout'
 export * from './hooks/useFormSpy'
 export * from './hooks/useFieldState'
 export * from './hooks/useFormState'
+export * from './hooks/useLayout'
 export * from './hooks/useForm'
 export * from './hooks/useField'
 export * from './hooks/useVirtualField'
